Add files via upload
"guns.csv" is the final version of the clean data used in my final project. The data was cleaned using an edited version of the "CDC_parser" function; the function used to clean and merge the data is in this github repository. This data set is a conglomerate of 7 data sets from the CDC's multiple casualty report. Each of the 7 data sets represents a different year between 2012 and 2018.